+++ build.gradle
@@ -21,7 +21,7 @@ repositories {
     }
 }
 
-version = "2.1.2"
+version = "2.1.1"
 group= "com.zerofall.ezstorage2"
 archivesBaseName = "ezstorage"
