+++ src/main/java/com/zerofall/ezstorage/network/EZNetwork.java
@@ -9,6 +9,7 @@ import net.minecraftforge.fml.common.network.NetworkRegistry.TargetPoint;
 import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import net.minecraftforge.fml.relauncher.Side;
 
+import com.zerofall.ezstorage.EZStorage;
 import com.zerofall.ezstorage.tileentity.TileEntitySecurityBox.SecurePlayer;
 
 public class EZNetwork {
@@ -20,9 +21,8 @@ public class EZNetwork {
 	public static SimpleNetworkWrapper registerNetwork() {
 		id = 0;
 		net = NetworkRegistry.INSTANCE.newSimpleChannel("ezChannel");
-    	net.registerMessage(MessageCustomClick.Handler.class, MessageCustomClick.class, id++, Side.SERVER);
-    	net.registerMessage(MessageRecipeSync.Handler.class, MessageRecipeSync.class, id++, Side.SERVER);
-    	net.registerMessage(MessageCraftingSync.Handler.class, MessageCraftingSync.class, id++, Side.CLIENT);
+    	net.registerMessage(PacketHandler.class, MyMessage.class, id++, Side.SERVER);
+    	net.registerMessage(RecipePacketHandler.class, RecipeMessage.class, id++, Side.SERVER);
 		net.registerMessage(MessageSecureSync.Handler.class, MessageSecureSync.class, id++, Side.CLIENT);
 		net.registerMessage(MessageSecurePlayer.Handler.class, MessageSecurePlayer.class, id++, Side.SERVER);
 		return net;
