--- src/main/java/com/zerofall/ezstorage/network/MessageCustomClick.java
@@ -1,59 +1,0 @@
-package com.zerofall.ezstorage.network;
-
-import com.zerofall.ezstorage.gui.server.ContainerStorageCore;
-
-import io.netty.buffer.ByteBuf;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.Container;
-import net.minecraftforge.fml.common.network.ByteBufUtils;
-import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
-import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
-import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
-
-/** Custom click message */
-public class MessageCustomClick implements IMessage {
-    
-    private int index;
-    private int button;
-    private int mode;
-
-    public MessageCustomClick() {}
-
-    public MessageCustomClick(int index, int button, int mode) {
-        this.index = index;
-        this.button = button;
-        this.mode = mode;
-    }
-
-    @Override
-    public void fromBytes(ByteBuf buf) {
-        index = ByteBufUtils.readVarInt(buf, 5);
-        button = ByteBufUtils.readVarInt(buf, 5);
-        mode = ByteBufUtils.readVarInt(buf, 5);
-    }
-
-    @Override
-    public void toBytes(ByteBuf buf) {
-    	ByteBufUtils.writeVarInt(buf, index, 5);
-        ByteBufUtils.writeVarInt(buf, button, 5);
-        ByteBufUtils.writeVarInt(buf, mode, 5);
-    }
-    
-    /** Handle the serverside custom slot click */
-    public static class Handler implements IMessageHandler<MessageCustomClick, IMessage> {
-        
-        @Override
-        public IMessage onMessage(MessageCustomClick message, MessageContext ctx) {
-        	EntityPlayer player = ctx.getServerHandler().playerEntity;
-        	Container container = player.openContainer;
-        	if (container != null && container instanceof ContainerStorageCore) {
-        		ContainerStorageCore storageContainer = (ContainerStorageCore)container;
-        		storageContainer.customSlotClick(message.index, message.button, message.mode, player);
-        	}
-        	return null; // no response in this case
-        }
-        
-        
-    }
-    
-}
