+++ src/main/java/com/zerofall/ezstorage/network/PacketHandler.java
@@ -1,0 +1,25 @@
+package com.zerofall.ezstorage.network;
+
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.inventory.Container;
+import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
+import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
+import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
+
+import com.zerofall.ezstorage.gui.server.ContainerStorageCore;
+
+public class PacketHandler implements IMessageHandler<MyMessage, IMessage> {
+    
+    @Override
+    public IMessage onMessage(MyMessage message, MessageContext ctx) {
+    	EntityPlayer player = ctx.getServerHandler().playerEntity;
+    	Container container = player.openContainer;
+    	if (container != null && container instanceof ContainerStorageCore) {
+    		ContainerStorageCore storageContainer = (ContainerStorageCore)container;
+    		storageContainer.customSlotClick(message.index, message.button, message.mode, player);
+    	}
+    	return null; // no response in this case
+    }
+    
+    
+}
