+++ src/main/java/com/zerofall/ezstorage/ref/RefStrings.java
@@ -3,7 +3,7 @@ package com.zerofall.ezstorage.ref;
 public class RefStrings {
 	public static final String MODID = "ezstorage";
 	public static final String NAME = "EZ Storage 2";
-	public static final String VERSION = "2.1.2";
+	public static final String VERSION = "2.1.1";
 	public static final String CLIENT_PROXY = "com.zerofall.ezstorage.proxy.ClientProxy";
 	public static final String SERVER_PROXY = "com.zerofall.ezstorage.proxy.CommonProxy";
 }
